requirements.txt
Flask
Flask-HTTPAuth
Flask-SQLAlchemy
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
itsdangerous
passlib
pyyaml
pytube
